migrate functional to the core app
