+++ analysis/tests/tests_forms.py
@@ -7,7 +7,6 @@ from dry_tests.models import Fields, TrueForm
 from analysis.forms import (
     OneTextForm,
     TwoTextForm,
-    LoadTrainingDataForm,
 )
 
 
@@ -54,28 +53,4 @@ class TestTwoTextForm(SimpleTestCase):
         )
 
         current_form = TwoTextForm()
-        self.assertTrueForm(current_form, true_form)
-
-
-class LoadTrainingDataFormSimpleTestCase(SimpleTestCase):
-    """
-    Load traning data test
-    """
-
-    def test_fields(self):
-        """
-        Test available fields
-        """
-        true_form = TrueForm(
-            fields=Fields(
-                count=3,
-                types={
-                    'name': forms.CharField,
-                    'excel_file': forms.FileField,
-                    'sheet_name': forms.IntegerField,
-                }
-            )
-        )
-
-        current_form = LoadTrainingDataForm()
         self.assertTrueForm(current_form, true_form)
