+++ core/templates/core/clear_data.html
