+++ core/templates/core/training_data.html
@@ -10,7 +10,7 @@
     </tr>
     <tr>
         <td>
-            <a href="{% url 'analysis:training_data' pk=object.pk %}">{{object.name}}</a>
+            <a href="{% url 'core:training_data' pk=object.pk %}">{{object.name}}</a>
         </td>
         <td>
             {{object.columns_count}}
@@ -22,10 +22,7 @@
 
 </table>
 <a class="btn btn-primary" href="{% url 'analysis:total_rating_form' pk=object.pk %}">Total rating</a>
-<a class="btn btn-info" href="#">One column rating</a>
-<a class="btn btn-info" href="{% url 'analysis:find_similar' pk=object.pk %}">Find similar</a>
-<a class="btn btn-second" href="{% url 'analysis:tokenize' %}">Tokenize</a>
-<a class="btn btn-danger" href="{% url 'analysis:delete_training_data' pk=object.pk %}">Delete</a>
+<a class="btn btn-danger" href="{% url 'core:delete_training_data' pk=object.pk %}">Delete</a>
 
 {% endblock %}
 {% block results %}
