+++ core/templates/core/training_data_delete_confirm.html
