+++ templates/base.html
@@ -49,6 +49,42 @@
   <div class="row">
     <nav id="sidebarMenu" class="col-md-3 col-lg-2 d-md-block bg-light sidebar collapse">
       <div class="position-sticky pt-3">
+         <h6 class="sidebar-heading d-flex justify-content-between align-items-center px-3 mt-4 mb-1 text-muted">
+          <span>Proximity Calculation</span>
+          <a class="link-secondary" href="#" aria-label="Add a new report">
+            <span data-feather="plus-circle"></span>
+          </a>
+        </h6>
+        <ul class="nav flex-column mb-2">
+         <li class="nav-item">
+            <a class="nav-link" href="{% url 'core:load_training_data' %}">
+              <span data-feather="file"></span>
+              Load Training Data
+            </a>
+          </li>
+           <li class="nav-item">
+            <a class="nav-link" href="{% url 'core:training_data_list' %}">
+              <span data-feather="file"></span>
+              Training Data List
+            </a>
+          </li>
+        </ul>
+
+        <h6 class="sidebar-heading d-flex justify-content-between align-items-center px-3 mt-4 mb-1 text-muted">
+          <span>Results</span>
+          <a class="link-secondary" href="#" aria-label="Add a new report">
+            <span data-feather="plus-circle"></span>
+          </a>
+        </h6>
+        <ul class="nav flex-column mb-2">
+         <li class="nav-item">
+            <a class="nav-link" href="{% url 'analysis:result_list' %}">
+              <span data-feather="file-text"></span>
+              Result List
+            </a>
+          </li>
+        </ul>
+
         <ul class="nav flex-column">
           <h6 class="sidebar-heading d-flex justify-content-between align-items-center px-3 mt-4 mb-1 text-muted">
           <span>Analysis</span>
@@ -66,6 +102,24 @@
             <a class="nav-link" href="{% url 'analysis:compare_two' %}">
               <span data-feather="file"></span>
               Compare two texts
+            </a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" href="{% url 'analysis:text_token_list' %}">
+              <span data-feather="file"></span>
+              TextToken List
+            </a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" href="{% url 'analysis:tokenize' %}">
+              <span data-feather="file"></span>
+              Tokenize
+            </a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" href="{% url 'analysis:find_similar' %}">
+              <span data-feather="file"></span>
+              Find Similar
             </a>
           </li>
         </ul>
@@ -91,46 +145,6 @@
           </li>
         </ul>
 
-        <h6 class="sidebar-heading d-flex justify-content-between align-items-center px-3 mt-4 mb-1 text-muted">
-          <span>Results</span>
-          <a class="link-secondary" href="#" aria-label="Add a new report">
-            <span data-feather="plus-circle"></span>
-          </a>
-        </h6>
-        <ul class="nav flex-column mb-2">
-         <li class="nav-item">
-            <a class="nav-link" href="{% url 'analysis:result_list' %}">
-              <span data-feather="file-text"></span>
-              Result List
-            </a>
-          </li>
-        </ul>
-        <h6 class="sidebar-heading d-flex justify-content-between align-items-center px-3 mt-4 mb-1 text-muted">
-          <span>Training data</span>
-          <a class="link-secondary" href="#" aria-label="Add a new report">
-            <span data-feather="plus-circle"></span>
-          </a>
-        </h6>
-        <ul class="nav flex-column mb-2">
-         <li class="nav-item">
-            <a class="nav-link" href="{% url 'analysis:load_training_data' %}">
-              <span data-feather="file"></span>
-              Load Training Data
-            </a>
-          </li>
-           <li class="nav-item">
-            <a class="nav-link" href="{% url 'analysis:training_data_list' %}">
-              <span data-feather="file"></span>
-              Training Data List
-            </a>
-          </li>
-           <li class="nav-item">
-            <a class="nav-link" href="{% url 'analysis:text_token_list' %}">
-              <span data-feather="file"></span>
-              TextToken List
-            </a>
-          </li>
-        </ul>
          <h6 class="sidebar-heading d-flex justify-content-between align-items-center px-3 mt-4 mb-1 text-muted">
           <span>Admin</span>
           <a class="link-secondary" href="#" aria-label="Add a new report">
